requirements.txt
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.4.1.post1
pyyaml
matplotlib
flask==3.0.2
uvicorn 
pydantic 
boto3==1.34.34
SQLAlchemy==2.0.27
psycopg2-binary